test(python): Add `glimpse` test
